refactor: clean architecture folder pattern
