+++ 04-clean-ddd/src/domain/forum/application/use-cases/answer-question.spec.ts
@@ -1,5 +1,5 @@
 import { AnswerQuestionUseCase } from "./answer-question";
-import { Answer } from "../entities/answer";
+import { Answer } from "../../enterprise/entities/answer";
 import { AnswersRepository } from "../repositories/answers-repository";
 
 const fakeAnswersRepository: AnswersRepository = {
