+++ 04-clean-ddd/src/domain/forum/application/use-cases/answer-question.ts
@@ -1,5 +1,5 @@
 import { UniqueEtityId } from "@/core/entities/unique-entity-id";
-import { Answer } from "../entities/answer";
+import { Answer } from "../../enterprise/entities/answer";
 import { AnswersRepository } from "../repositories/answers-repository";
 
 interface AnswerQuestionUseCaseRequest {
