+++ 04-clean-ddd/src/domain/forum/enterprise/entities/student.ts
